studynotion-edtech-project-main/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.1
pydantic-settings==2.5.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cloudinary==1.41.0
razorpay==1.4.2
python-dotenv==1.0.1
